Restore lineRanges.endX value - ?KateDocument behaviour changed?
Fix MakeVisible:
- non-wrapped mode: account for scrollbar appearing during scroll
  - still needs a bit more tweaking

- wrapped mode: account for wrapped lines
  - almost there, just some cursor down issues (cursor disappears
  off the edge 50% of time) and the end-of-document issue (can't scroll
  to end when last few lines are wrapped yet)

Fix PageUp / PageDown in wrapped mode:
- accout for wrapped lines

Fix endLine() in wrapped mode:
- account for wrapped lines

Added some documentation to functions to remind myself of what they do.

A few uint <-> int comparison fixes.

There's quite a bit of cruft is this commit (obsolete code, heaps of kdDebug()s)
which I will clean out soon.  I'm putting it in now because it's better than
what is already there :)  (the non-wrapped code paths haven't changed
barring the scrollbar appearing tweak)

svn path=/trunk/kdelibs/kate/; revision=179243

+++ part/part/kateiconborder.cpp
@@ -245,7 +245,7 @@ void KateIconBorder::paintBorder (int x, int y, int width, int height)
   int w( this->width() );                     // sane value/calc only once
   int lnbx( 2+lnWidth-1 );              // line nbr pane border position: calc only once 
 
-  int currentLine( m_viewInternal->cursor.line );
+  int currentLine( (int)m_viewInternal->cursor.line );
   
   QPainter p ( this );
   p.setFont ( m_doc->getFont(KateDocument::ViewFont) ); // for line numbers
